refactor: divide code in issue view controller
- 확장 코드를 다른 파일로 분리시킴
