+++ iOS/IssueTracker/IssueTracker/Issue/IssueList/Controller/IssueViewController+SearchUpdate.swift
@@ -1,0 +1,15 @@
+//
+//  IssueViewController+SearchUpdate.swift
+//  IssueTracker
+//
+//  Created by 채훈기 on 2020/11/03.
+//
+
+import UIKit
+
+extension IssueViewController: UISearchResultsUpdating {
+    func updateSearchResults(for searchController: UISearchController) {
+        guard let text = searchController.searchBar.text else { return }
+        dataSource.performQuery(issues: viewModel.issues, filter: TitleCriteria(input: text))
+    }
+}
