--- iOS/IssueTracker/IssueTracker/Issue/IssueNavigationViewController.swift
@@ -1,15 +1,0 @@
-//
-//  IssueNavigationViewController.swift
-//  IssueTracker
-//
-//  Created by 채훈기 on 2020/10/27.
-//
-
-import UIKit
-
-class IssueNavigationViewController: UINavigationController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
